Refactor functions that get time, date and day to be sync
Create a setTimeZone function that stores current timeZone in a variable
and that variable is only changed when location is changed.

This lets you only have to call the api once everytime you change
the location to get its time zone instead of calling it everytime the
time the time changes.