Refactor the model to only call the api less
Make the function that gets makes the api call store the json it
receives. Then the rest of the functions in the code that get info such
as current temperature from the api, would start getting it from the
json stored from that original api call.

The variable containing the forecast json only updates when the api is
called. The api is called every hour if the user doesn't change their
location and whenever the user looks up a new location.